Fix All toggle render
The All toggle doesn't render correctly due to this issue --> https://github.com/mui-org/material-ui/issues/18018. I've set the All toggle to render only when there are more than one reporters. This fixes the visual render for now but leaves the warning about the className mismatch between server and the client in the console.